requirements.txt


djangorestframework
markdown      
django-filter 
django-cors-headers
pyotp
djangorestframework-simplejwt
